out_stackdriver: merge insertId/operation/sourceLocation into one file
