+++ plugins/out_stackdriver/CMakeLists.txt
@@ -2,8 +2,7 @@ set(src
   gce_metadata.c
   stackdriver_conf.c
   stackdriver.c
-  stackdriver_operation.c
-  stackdriver_sourceLocation.c
+  stackdriver_special_fields.c
   )
 
 FLB_PLUGIN(out_stackdriver "${src}" "")
