+++ tests/runtime/CMakeLists.txt
@@ -56,7 +56,6 @@ if(FLB_IN_LIB)
   FLB_RT_TEST(FLB_OUT_RETRY        "out_retry.c")
   FLB_RT_TEST(FLB_OUT_STDOUT       "out_stdout.c")
   FLB_RT_TEST(FLB_OUT_TD           "out_td.c")
-  FLB_RT_TEST(FLB_OUT_STACKDRIVER  "out_stackdriver.c")
 
 endif()
